+++ src/test/java/roomescape/MissionStep1Test.java
@@ -1,0 +1,55 @@
+package roomescape;
+
+import io.restassured.RestAssured;
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.annotation.DirtiesContext;
+
+import static org.hamcrest.Matchers.is;
+
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
+@DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
+public class MissionStep1Test {
+
+    @Test
+    void 예약_조회() {
+        RestAssured.given().log().all()
+                .when().get("/reservations")
+                .then().log().all()
+                .statusCode(200)
+                .body("size()", is(0)); // 아직 생성 요청이 없으니 0개
+    }
+
+//    @Test
+//    void 예약_추가_및_삭제() {
+//        Map<String, String> params = new HashMap<>();
+//        params.put("name", "브라운");
+//        params.put("date", "2023-08-05");
+//        params.put("time", "15:40");
+//
+//        RestAssured.given().log().all()
+//                .contentType(ContentType.JSON)
+//                .body(params)
+//                .when().post("/reservations")
+//                .then().log().all()
+//                .statusCode(200)
+//                .body("id", is(1));
+//
+//        RestAssured.given().log().all()
+//                .when().get("/reservations")
+//                .then().log().all()
+//                .statusCode(200)
+//                .body("size()", is(1));
+//
+//        RestAssured.given().log().all()
+//                .when().delete("/reservations/1")
+//                .then().log().all()
+//                .statusCode(200);
+//
+//        RestAssured.given().log().all()
+//                .when().get("/reservations")
+//                .then().log().all()
+//                .statusCode(200)
+//                .body("size()", is(0));
+//    }
+}
